requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
Django==4.2.4
django-graphql-jwt==0.4.0
graphene==3.3
graphene-django==3.1.3
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
PyJWT==2.8.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
